+++ Twilio/Rest/Api/V2010/Account/IncomingPhoneNumberList.php
@@ -15,6 +15,7 @@ use Twilio\Options;
 use Twilio\Rest\Api\V2010\Account\IncomingPhoneNumber\LocalList;
 use Twilio\Rest\Api\V2010\Account\IncomingPhoneNumber\MobileList;
 use Twilio\Rest\Api\V2010\Account\IncomingPhoneNumber\TollFreeList;
+use Twilio\Serialize;
 use Twilio\Values;
 use Twilio\Version;
 
@@ -108,7 +109,7 @@ class IncomingPhoneNumberList extends ListResource {
     public function page($options = array(), $pageSize = Values::NONE, $pageToken = Values::NONE, $pageNumber = Values::NONE) {
         $options = new Values($options);
         $params = Values::of(array(
-            'Beta' => $options['beta'],
+            'Beta' => Serialize::booleanToString($options['beta']),
             'FriendlyName' => $options['friendlyName'],
             'PhoneNumber' => $options['phoneNumber'],
             'PageToken' => $pageToken,
@@ -147,7 +148,7 @@ class IncomingPhoneNumberList extends ListResource {
             'StatusCallback' => $options['statusCallback'],
             'StatusCallbackMethod' => $options['statusCallbackMethod'],
             'VoiceApplicationSid' => $options['voiceApplicationSid'],
-            'VoiceCallerIdLookup' => $options['voiceCallerIdLookup'],
+            'VoiceCallerIdLookup' => Serialize::booleanToString($options['voiceCallerIdLookup']),
             'VoiceFallbackMethod' => $options['voiceFallbackMethod'],
             'VoiceFallbackUrl' => $options['voiceFallbackUrl'],
             'VoiceMethod' => $options['voiceMethod'],