+++ CHANGES.md
@@ -17,6 +17,7 @@ Breaking Changes, refer to [Upgrade Guide][upgrade]
     - Credential List Mappings, IP ACL Mappings, SIP Domains
  - Fix incorrect types documentation of `links`/`subresourceUri` fields on various resources. Was incorrectly documented as string, actual type was an array.
  - Fix some properties incorrectly documented as `string` when actually were `array` types.
+ - Fix boolean parameters did not accept boolean values, now accept both boolean and strings for backwards compatibility.
  - Remove `uri` field from Pricing Phone Number Countries resource (backwards incompatible)
  - Properly deserialize date times for various resources (backwards incompatible)
  - Accept DateTime inputs for date parameters for various resources (backwards incompatible)